requirements.txt
fastapi>=0.115.0
mcp[cli]>=1.2.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.28.0
stripe>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
